+++ tryit-frontend/components/index.ts
@@ -1,6 +1,6 @@
 export { default as Map } from "./home/Map.vue"
 export { default as Timer } from "./home/Timer.vue"
-export { default as PopupForm } from "./home/PopupForm.vue"
+export { default as PopupTicket } from "./home/PopupTicket.vue"
 export { default as PopupLogin } from "./home/PopupLogin.vue"
 export { default as DialButton } from "./home/DialButton.vue"
 export { default as Timeline } from "./home/Timeline.vue"
